refactor(server): split routes.rs into per-endpoint submodules (ENG-1747 phase 4.2)
routes.rs (2851 lines) → src/routes/{mod,remember,recall,analyze,admin,sponsor}.rs.
Pure code move — no behaviour change. Each handler keeps its body verbatim;
shared route-level helpers (enqueue_wallet_job, truncate_str,
collect_bounded_results, cleanup_expired_blob) live in routes/mod.rs, which
re-exports every handler so main.rs keeps using routes::<name> unchanged.

Module map:
  - remember.rs — /api/remember{,/manual,/bulk,/bulk/status} + the async prep
    tasks (spawn_prepare_*) + the summarize-for-embedding helpers (ENG-1407)
  - recall.rs   — /api/recall{,/manual} + the recall query-embedding cache (ENG-1405)
  - analyze.rs  — /api/analyze (production async path + benchmark-mode sync bypass)
  - admin.rs    — /api/ask, /api/forget, /api/stats, /api/restore, /health, /config
  - sponsor.rs  — /sponsor, /sponsor/execute (Enoki proxy)
  - mod.rs      — shared helpers + handler re-exports

Tests moved with their code (135 total, all green); the ASK_SYSTEM_PROMPT
include_str! path changed services/prompts/ → ../services/prompts/. No new
clippy warnings.

Refs ENG-1747
