+++ services/server/src/routes/admin.rs
@@ -1,0 +1,614 @@
+//! Admin / utility handlers: `/api/ask`, `/api/forget`, `/api/stats`,
+//! `/api/restore`, `GET /health`, `GET /config`.
+//!
+//! `ask` is the AI-with-memory demo (recall → inject memories into the LLM
+//! system prompt → answer). `forget`/`stats` are owner-scoped, mode-blind
+//! admin ops the benchmark harness uses for cleanup/verification. `restore`
+//! rebuilds a namespace's vector index from the on-chain blobs (download →
+//! SEAL-decrypt → re-embed → insert missing rows). `/health` reports the
+//! deployment mode; `/config` exposes the public Sui/package metadata the
+//! SDK needs to build a SEAL SessionKey.
+
+use axum::extract::State;
+use axum::{Extension, Json};
+use futures::stream::{self, StreamExt};
+use std::sync::Arc;
+
+use crate::services::llm_chat::{ChatCompletionRequest, ChatCompletionResponse, ChatMessage};
+use crate::storage::{seal, walrus};
+use crate::types::*;
+
+use super::{cleanup_expired_blob, truncate_str};
+
+/// ENG-1747: the `/api/ask` system prompt — a versioned text asset with a
+/// `{MEMORY_CONTEXT}` placeholder (substituted with the `<memory>`-tag-
+/// wrapped recall context per request). Includes the LOW-8 prompt-injection
+/// guard. Bundled at compile time.
+const ASK_SYSTEM_PROMPT: &str = include_str!("../services/prompts/ask.txt");
+/// Version ID for the ask prompt. Bump on every meaningful prompt change.
+#[allow(dead_code)]
+const ASK_SYSTEM_PROMPT_VERSION: &str = "ask.v1";
+
+// ============================================================
+// /api/forget + /api/stats
+// ============================================================
+
+/// POST /api/forget
+///
+/// Delete the vector index rows for every memory in `owner`'s
+/// `namespace` (a hard `DELETE` on `vector_entries` — the underlying
+/// Walrus blobs persist, since Walrus has no delete; the memories just
+/// stop being retrievable and stop counting toward storage quota). Used
+/// by the benchmark harness for inter-run cleanup; also a general admin
+/// op. Mode-blind — works the same in production and benchmark mode (in
+/// benchmark mode this also removes the plaintext rows). Owner-scoped:
+/// only the caller's own rows are deleted.
+pub async fn forget(
+    State(state): State<Arc<AppState>>,
+    Extension(auth): Extension<AuthInfo>,
+    Json(body): Json<ForgetRequest>,
+) -> Result<Json<ForgetResponse>, AppError> {
+    let owner = &auth.owner;
+    let namespace = &body.namespace;
+    tracing::info!("forget: owner={} ns={}", owner, namespace);
+
+    let deleted = state.db.delete_by_namespace(owner, namespace).await?;
+
+    tracing::info!(
+        "forget complete: deleted {} entries for owner={} ns={}",
+        deleted,
+        owner,
+        namespace
+    );
+
+    Ok(Json(ForgetResponse {
+        deleted,
+        namespace: namespace.clone(),
+        owner: owner.clone(),
+    }))
+}
+
+/// POST /api/stats
+///
+/// Return memory count + stored bytes for `owner`'s `namespace`. Used by
+/// the benchmark harness to verify ingestion. Mode-blind. Owner-scoped.
+pub async fn stats(
+    State(state): State<Arc<AppState>>,
+    Extension(auth): Extension<AuthInfo>,
+    Json(body): Json<StatsRequest>,
+) -> Result<Json<StatsResponse>, AppError> {
+    let owner = &auth.owner;
+    let namespace = &body.namespace;
+
+    let (memory_count, storage_bytes) = state.db.namespace_stats(owner, namespace).await?;
+
+    tracing::info!(
+        "stats: owner={} ns={} count={} bytes={}",
+        owner,
+        namespace,
+        memory_count,
+        storage_bytes
+    );
+
+    Ok(Json(StatsResponse {
+        memory_count,
+        storage_bytes,
+        namespace: namespace.clone(),
+        owner: owner.clone(),
+    }))
+}
+
+// ============================================================
+// /health + /config
+// ============================================================
+
+/// GET /health
+pub async fn health(State(state): State<Arc<AppState>>) -> Json<HealthResponse> {
+    Json(HealthResponse {
+        status: "ok".to_string(),
+        version: env!("CARGO_PKG_VERSION").to_string(),
+        mode: if state.config.benchmark_mode {
+            "benchmark".to_string()
+        } else {
+            "production".to_string()
+        },
+    })
+}
+
+/// GET /config
+///
+/// ENG-1697: public, unauthenticated endpoint returning deployment
+/// parameters the SDK needs to build a SEAL `SessionKey` client-side —
+/// specifically the Move `packageId` and the Sui network/RPC URL.
+///
+/// These values are public on-chain metadata (not secrets), so no auth is
+/// required. Exposing them here lets the SDK migrate from transmitting
+/// the raw delegate private key (`x-delegate-key`) to transmitting an
+/// exported SessionKey (`x-seal-session`) without forcing users to add
+/// `packageId` to their `MemWalConfig` — preserving backward-compatible
+/// UX for v0.3.x apps that only passed `{ key, accountId }`.
+pub async fn get_config(State(state): State<Arc<AppState>>) -> Json<ConfigResponse> {
+    Json(ConfigResponse {
+        package_id: state.config.package_id.clone(),
+        network: state.config.sui_network.clone(),
+        sui_rpc_url: state.config.sui_rpc_url.clone(),
+        rate_limit_disabled: state.config.rate_limit.bench_bypass_enabled,
+    })
+}
+
+// ============================================================
+// /api/ask
+// ============================================================
+
+/// POST /api/ask
+///
+/// Full AI-with-memory demo:
+/// 1. Recall relevant memories for the question
+/// 2. Inject memories into LLM system prompt
+/// 3. Call LLM with user question + memory context
+/// 4. Return answer + memories used
+pub async fn ask(
+    State(state): State<Arc<AppState>>,
+    Extension(auth): Extension<AuthInfo>,
+    Json(body): Json<AskRequest>,
+) -> Result<Json<AskResponse>, AppError> {
+    if body.question.is_empty() {
+        return Err(AppError::BadRequest("Question cannot be empty".into()));
+    }
+
+    let owner = &auth.owner;
+    let namespace = &body.namespace;
+    let limit = body.limit.unwrap_or(5);
+    tracing::info!(
+        "ask: question=\"{}...\" owner={} ns={}",
+        truncate_str(&body.question, 50),
+        owner,
+        namespace
+    );
+
+    // Step 1: Recall relevant memories
+    let query_vector = state.embedder.embed(&body.question).await?;
+    let hits = state
+        .db
+        .search_similar(&query_vector, owner, namespace, limit)
+        .await?;
+
+    // Hydrate the hits through the storage engine, concurrently — same
+    // blob cache -> Walrus download -> SEAL decrypt -> UTF-8 path as
+    // recall, with reactive cleanup on Walrus 404. The engine derives the
+    // SEAL credential from `auth`; per-blob errors are logged inside it.
+    let fetch_tasks = hits.into_iter().map(|hit| {
+        let auth = &auth;
+        let engine = &state.engine;
+        async move { engine.fetch_one(owner, &hit.blob_id, hit.distance, auth).await }
+    });
+    let memories: Vec<RecallResult> = futures::future::join_all(fetch_tasks)
+        .await
+        .into_iter()
+        // fetch_one returns Ok(None) for blobs that are gone / failed to
+        // decrypt; surface only the AppError (sidecar down, etc.).
+        .filter_map(|r| match r {
+            Ok(Some(m)) => Some(Ok(RecallResult {
+                blob_id: m.blob_id,
+                text: m.text,
+                distance: m.distance,
+            })),
+            Ok(None) => None,
+            Err(e) => Some(Err(e)),
+        })
+        .collect::<Result<Vec<_>, AppError>>()?;
+
+    let memories_used = memories.len();
+    tracing::info!("ask: {} memories found for context", memories_used);
+
+    // LOW-8: Defence-in-depth against indirect prompt injection via stored memories.
+    // Wrap each memory in an explicit <memory> tag with the blob_id and tell the
+    // LLM in the system prompt that tag contents are user-provided data, not
+    // instructions. This does not eliminate the attack vector (owner-scoped
+    // memories can still contain adversarial text) but makes tag-boundary
+    // confusion attacks harder to mount.
+    let memory_context = if memories.is_empty() {
+        "No memories found for this user yet.".to_string()
+    } else {
+        let lines: Vec<String> = memories
+            .iter()
+            .map(|m| {
+                format!(
+                    "<memory id=\"{}\" relevance=\"{:.2}\">{}</memory>",
+                    m.blob_id,
+                    1.0 - m.distance,
+                    m.text
+                )
+            })
+            .collect();
+        format!("Known facts about this user:\n{}", lines.join("\n"))
+    };
+
+    // ENG-1747: the ask system prompt is a versioned text asset
+    // (services/prompts/ask.txt) with a {MEMORY_CONTEXT} placeholder.
+    // Keeps the LOW-8 prompt-injection guard. ASK_SYSTEM_PROMPT_VERSION
+    // tracks the prompt version for attribution.
+    let system_prompt = ASK_SYSTEM_PROMPT.replace("{MEMORY_CONTEXT}", &memory_context);
+
+    // Step 3: Call LLM
+    let api_key = state
+        .config
+        .openai_api_key
+        .as_ref()
+        .ok_or_else(|| AppError::Internal("OPENAI_API_KEY required for /api/ask".into()))?;
+    let url = format!("{}/chat/completions", state.config.openai_api_base);
+
+    let resp = state
+        .http_client
+        .post(&url)
+        .header("Authorization", format!("Bearer {}", api_key))
+        .header("Content-Type", "application/json")
+        .json(&ChatCompletionRequest {
+            model: "openai/gpt-4o-mini".to_string(),
+            messages: vec![
+                ChatMessage {
+                    role: "system".to_string(),
+                    content: system_prompt,
+                },
+                ChatMessage {
+                    role: "user".to_string(),
+                    content: body.question.clone(),
+                },
+            ],
+            temperature: 0.7,
+            max_tokens: 512,
+        })
+        .send()
+        .await
+        .map_err(|e| AppError::Internal(format!("LLM request failed: {}", e)))?;
+
+    if !resp.status().is_success() {
+        let status = resp.status();
+        let body_text = resp.text().await.unwrap_or_default();
+        return Err(AppError::Internal(format!(
+            "LLM error ({}): {}",
+            status, body_text
+        )));
+    }
+
+    let api_resp: ChatCompletionResponse = resp
+        .json()
+        .await
+        .map_err(|e| AppError::Internal(format!("Failed to parse LLM response: {}", e)))?;
+
+    let answer = api_resp
+        .choices
+        .first()
+        .map(|c| c.message.content.trim().to_string())
+        .unwrap_or_else(|| "No response from LLM".to_string());
+
+    tracing::info!("ask complete: answer length={} chars", answer.len());
+
+    Ok(Json(AskResponse {
+        answer,
+        memories_used,
+        memories,
+    }))
+}
+
+// ============================================================
+// /api/restore
+// ============================================================
+
+/// POST /api/restore
+///
+/// Restore a namespace from Walrus:
+/// 1. Get all blob_ids for owner+namespace from DB
+/// 2. Download each blob from Walrus
+/// 3. SEAL decrypt with delegate key
+/// 4. Re-embed decrypted text
+/// 5. Clear old vector entries and re-index
+pub async fn restore(
+    State(state): State<Arc<AppState>>,
+    Extension(auth): Extension<AuthInfo>,
+    Json(body): Json<RestoreRequest>,
+) -> Result<Json<RestoreResponse>, AppError> {
+    if body.namespace.is_empty() {
+        return Err(AppError::BadRequest("namespace cannot be empty".into()));
+    }
+
+    let owner = &auth.owner;
+    let namespace = &body.namespace;
+    let limit = body.limit;
+    tracing::info!("restore: owner={} ns={} limit={}", owner, namespace, limit);
+
+    // ENG-1697: Prefer the client-built SessionKey; fall back to legacy
+    // delegate key, then to the server's own key for restore operations.
+    let credential = seal::SealCredential::from_auth_or_fallback(
+        &auth,
+        state.config.sui_private_key.as_deref(),
+    )
+    .ok_or_else(|| {
+        AppError::Internal(
+            "SEAL credential required for restore (x-seal-session, x-delegate-key, or SERVER_SUI_PRIVATE_KEY)".into(),
+        )
+    })?;
+
+    // Step 1: Discover all blob_ids from on-chain (source of truth)
+    tracing::info!(
+        "restore: querying chain for blobs owner={} ns={}",
+        owner,
+        namespace
+    );
+    let on_chain_blobs = walrus::query_blobs_by_owner(
+        &state.http_client,
+        &state.config.sidecar_url,
+        state.config.sidecar_secret.as_deref(),
+        owner,
+        Some(namespace),
+        Some(&state.config.package_id),
+    )
+    .await?;
+    let all_blob_ids: Vec<String> = on_chain_blobs.iter().map(|b| b.blob_id.clone()).collect();
+    let total = all_blob_ids.len();
+
+    // Build blob_id → package_id lookup from on-chain metadata
+    // Each blob may have been encrypted with a different package_id (e.g. after contract upgrades)
+    let blob_package_ids: std::collections::HashMap<String, String> = on_chain_blobs
+        .iter()
+        .filter(|b| !b.package_id.is_empty())
+        .map(|b| (b.blob_id.clone(), b.package_id.clone()))
+        .collect();
+
+    if total == 0 {
+        return Ok(Json(RestoreResponse {
+            restored: 0,
+            skipped: 0,
+            total: 0,
+            namespace: namespace.clone(),
+            owner: owner.clone(),
+        }));
+    }
+
+    // Step 2: Check which blobs already exist in local DB → only restore missing ones
+    let existing_blob_ids = state.db.get_blobs_by_namespace(owner, namespace).await?;
+    let existing_set: std::collections::HashSet<&str> =
+        existing_blob_ids.iter().map(|s| s.as_str()).collect();
+    let all_missing: Vec<String> = all_blob_ids
+        .iter()
+        .filter(|id| !existing_set.contains(id.as_str()))
+        .cloned()
+        .collect();
+    // Apply limit — take the most recent N missing blobs (last N from chain query)
+    let missing_blob_ids: Vec<String> = if all_missing.len() > limit {
+        all_missing[all_missing.len() - limit..].to_vec()
+    } else {
+        all_missing
+    };
+    let skipped = total - missing_blob_ids.len();
+    tracing::info!(
+        "restore: total={} on-chain, existing={}, missing={} (limited to {}) for ns={}",
+        total,
+        existing_blob_ids.len(),
+        missing_blob_ids.len(),
+        limit,
+        namespace
+    );
+
+    if missing_blob_ids.is_empty() {
+        return Ok(Json(RestoreResponse {
+            restored: 0,
+            skipped,
+            total,
+            namespace: namespace.clone(),
+            owner: owner.clone(),
+        }));
+    }
+
+    // Step 3: Download all missing blobs from Walrus concurrently
+    let db = &state.db;
+    let download_tasks: Vec<_> = missing_blob_ids
+        .iter()
+        .map(|blob_id| {
+            let walrus_client = &state.walrus_client;
+            let blob_id = blob_id.clone();
+            let owner_for_cleanup = owner.clone();
+            async move {
+                match walrus::download_blob(walrus_client, &blob_id).await {
+                    Ok(data) => Some((blob_id, data)),
+                    Err(AppError::BlobNotFound(msg)) => {
+                        tracing::warn!("restore: blob expired, skipping: {}", msg);
+                        cleanup_expired_blob(db, &blob_id, &owner_for_cleanup).await;
+                        None
+                    }
+                    Err(e) => {
+                        tracing::warn!("restore: download failed for {}: {}", blob_id, e);
+                        None
+                    }
+                }
+            }
+        })
+        .collect();
+
+    // MED-6 fix: Bounded concurrency (max 10 parallel downloads) to prevent
+    // OOM when restoring large namespaces. join_all() with hundreds of blobs
+    // would spawn all downloads simultaneously → memory spike.
+    // We use buffer_unordered(10) to cap parallelism at 10 concurrent downloads.
+    let downloaded: Vec<(String, Vec<u8>)> = stream::iter(download_tasks)
+        .buffer_unordered(10)
+        .filter_map(|opt| async move { opt })
+        .collect()
+        .await;
+
+    // Preserve encrypted blob sizes so restored rows still contribute to storage quota.
+    let blob_sizes: std::collections::HashMap<String, i64> = downloaded
+        .iter()
+        .map(|(blob_id, data)| (blob_id.clone(), data.len() as i64))
+        .collect();
+
+    if downloaded.is_empty() {
+        return Ok(Json(RestoreResponse {
+            restored: 0,
+            skipped,
+            total,
+            namespace: namespace.clone(),
+            owner: owner.clone(),
+        }));
+    }
+
+    tracing::info!(
+        "restore: downloaded {}/{} blobs, decrypting (3 concurrent)...",
+        downloaded.len(),
+        missing_blob_ids.len()
+    );
+
+    // Step 4: SEAL decrypt with bounded concurrency (3 at a time)
+    // Use per-blob package_id from on-chain metadata, fall back to current server config
+    let decrypt_results: Vec<Option<(String, String)>> = stream::iter(downloaded)
+        .map(|(blob_id, encrypted_data)| {
+            let http_client = &state.http_client;
+            let sidecar_url = state.config.sidecar_url.clone();
+            let sidecar_secret = state.config.sidecar_secret.clone();
+            let credential = credential.clone();
+            // Use the package_id that was stored with this blob (supports contract upgrades)
+            let package_id = blob_package_ids
+                .get(&blob_id)
+                .cloned()
+                .unwrap_or_else(|| state.config.package_id.clone());
+            let account_id = auth.account_id.clone();
+            async move {
+                match seal::seal_decrypt(
+                    http_client,
+                    &sidecar_url,
+                    sidecar_secret.as_deref(),
+                    &encrypted_data,
+                    &credential,
+                    &package_id,
+                    &account_id,
+                )
+                .await
+                {
+                    Ok(plaintext) => match String::from_utf8(plaintext) {
+                        Ok(text) => Some((blob_id, text)),
+                        Err(e) => {
+                            tracing::warn!("restore: invalid UTF-8 for {}: {}", blob_id, e);
+                            None
+                        }
+                    },
+                    Err(e) => {
+                        tracing::warn!("restore: decrypt failed for {}: {}", blob_id, e);
+                        None
+                    }
+                }
+            }
+        })
+        .buffer_unordered(3)
+        .collect()
+        .await;
+
+    let decrypted_texts: Vec<(String, String)> = decrypt_results.into_iter().flatten().collect();
+    tracing::info!(
+        "restore: decrypted {}/{} blobs",
+        decrypted_texts.len(),
+        missing_blob_ids.len()
+    );
+
+    // Step 5: Re-embed all decrypted texts concurrently
+    let embed_tasks: Vec<_> = decrypted_texts
+        .iter()
+        .map(|(blob_id, text)| {
+            let embedder = Arc::clone(&state.embedder);
+            let blob_id = blob_id.clone();
+            let text = text.clone();
+            async move {
+                match embedder.embed(&text).await {
+                    Ok(vector) => Some((blob_id, vector)),
+                    Err(e) => {
+                        tracing::warn!("restore: embedding failed for {}: {}", blob_id, e);
+                        None
+                    }
+                }
+            }
+        })
+        .collect();
+
+    let results: Vec<(String, Vec<f32>)> = futures::future::join_all(embed_tasks)
+        .await
+        .into_iter()
+        .flatten()
+        .collect();
+
+    // Step 6: Insert only new entries (no delete!)
+    let restored = results.len();
+    for (blob_id, vector) in &results {
+        let id = uuid::Uuid::new_v4().to_string();
+        let blob_size = blob_sizes.get(blob_id).copied().unwrap_or_else(|| {
+            tracing::warn!(
+                "restore: missing blob size for {}, defaulting to 0 for quota tracking",
+                blob_id
+            );
+            0
+        });
+        state
+            .db
+            .insert_vector(&id, owner, namespace, blob_id, vector, blob_size)
+            .await?;
+    }
+
+    tracing::info!(
+        "restore complete: restored={} skipped={} total={} owner={} ns={}",
+        restored,
+        skipped,
+        total,
+        owner,
+        namespace
+    );
+
+    Ok(Json(RestoreResponse {
+        restored,
+        skipped,
+        total,
+        namespace: namespace.clone(),
+        owner: owner.clone(),
+    }))
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::types::RecallResult;
+
+    // ── LOW-8: Memory context wraps in XML tags ─────────────────────────
+
+    #[test]
+    fn memory_context_uses_xml_tags() {
+        // Simulate what /api/ask does
+        let memories = [RecallResult {
+            blob_id: "blob123".into(),
+            text: "User likes coffee".into(),
+            distance: 0.1,
+        }];
+
+        let lines: Vec<String> = memories
+            .iter()
+            .map(|m| {
+                format!(
+                    "<memory id=\"{}\" relevance=\"{:.2}\">{}</memory>",
+                    m.blob_id,
+                    1.0 - m.distance,
+                    m.text
+                )
+            })
+            .collect();
+        let context = format!("Known facts about this user:\n{}", lines.join("\n"));
+
+        assert!(context.contains("<memory id=\"blob123\""));
+        assert!(context.contains("relevance=\"0.90\""));
+        assert!(context.contains("User likes coffee</memory>"));
+    }
+
+    #[test]
+    fn memory_context_empty_shows_no_memories() {
+        let memories: Vec<RecallResult> = vec![];
+        let context = if memories.is_empty() {
+            "No memories found for this user yet.".to_string()
+        } else {
+            "should not reach here".to_string()
+        };
+        assert_eq!(context, "No memories found for this user yet.");
+    }
+}
